pyproject.toml
[tool.poetry]
name = "service-scheduling"
version = "0.1.0"
description = ""
authors = ["Evandro Paulo Folletto <evandrofolletto@gmail.com>"]
readme = "README.md"
packages = [{include = "service_scheduling"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"
markdown = "^3.6"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
